+++ search_result.php
@@ -1,10 +1,9 @@
 <?php
 require_once "class/search.php";
-//require_once "class/mysql.php";
 require_once "class/page.php";
 
 
-$search_word = $_GET['search']; //検索ワード 
+$search_word = $_GET['search_word']; //検索ワード 
 $like_search = "%".$search_word."%";
 
 //GETメソッドで検索ワードを受けたら実行
